requirements.txt
matplotlib==3.8.4
numpy==2.4.3
pandas==3.0.1
Pillow==12.1.1
scikit_learn==1.4.2
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.66.4
